Rr friends and message

+++ app/src/main/java/northeastern/cs5520fa23/greenthumbs/viewmodel/Profile/ProfileFragment.java
@@ -2,17 +2,21 @@ package northeastern.cs5520fa23.greenthumbs.viewmodel.Profile;
 
 import static android.content.ContentValues.TAG;
 
+import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.drawable.BitmapDrawable;
 import android.net.Uri;
 import android.os.Bundle;
 
+import androidx.activity.OnBackPressedCallback;
 import androidx.activity.result.ActivityResultLauncher;
 import androidx.activity.result.PickVisualMediaRequest;
 import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
@@ -41,6 +45,7 @@ import com.google.firebase.database.ValueEventListener;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 import com.google.firebase.storage.UploadTask;
+import com.squareup.picasso.Picasso;
 
 import java.io.ByteArrayOutputStream;
 import java.util.ArrayList;
@@ -51,6 +56,7 @@ import java.util.Map;
 
 import northeastern.cs5520fa23.greenthumbs.MainActivity;
 import northeastern.cs5520fa23.greenthumbs.R;
+import northeastern.cs5520fa23.greenthumbs.viewmodel.FriendsUsers.UsersActivity;
 import northeastern.cs5520fa23.greenthumbs.viewmodel.SocialFeed.CreatePostDialog;
 import northeastern.cs5520fa23.greenthumbs.viewmodel.SocialFeed.ImgPost;
 import northeastern.cs5520fa23.greenthumbs.viewmodel.SocialFeed.SocialAdapter;
@@ -137,6 +143,22 @@ public class ProfileFragment extends Fragment {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
+        /*
+        OnBackPressedCallback backPressedCallback = new OnBackPressedCallback(true) {
+            @Override
+            public void handleOnBackPressed() {
+                FragmentManager manager = getActivity().getSupportFragmentManager();
+                manager.popBackStackImmediate();
+
+
+
+
+
+            }
+        };
+        getActivity().getOnBackPressedDispatcher().addCallback(getActivity(), backPressedCallback);
+        */
+
         this.usernameView = view.findViewById(R.id.profile_name);
         this.headerImage = view.findViewById(R.id.profile_header_image);
         this.profilePicture = view.findViewById(R.id.profile_prof_pic);
@@ -190,6 +212,8 @@ public class ProfileFragment extends Fragment {
             this.sendMsgButton.setVisibility(View.GONE);
             this.sendMsgButton.setEnabled(false);
             this.addFriendButton.setVisibility(View.GONE);
+            this.profilePicture.setEnabled(false);
+            this.headerImage.setEnabled(false);
             this.editProfileButton.setEnabled(true);
             this.editProfileButton.setVisibility(View.VISIBLE);
             this.editProfileButton.setOnClickListener(new View.OnClickListener() {
@@ -198,30 +222,33 @@ public class ProfileFragment extends Fragment {
                     if (!isEditing) {
                         isEditing = true;
                         editProfileButton.setText("Finish");
-                        usernameView.setEnabled(true);
+                        //usernameView.setEnabled(true);
                         userBioView.setEnabled(true);
+                        headerImage.setEnabled(true);
+                        profilePicture.setEnabled(true);
                     } else {
                         saveProfileUpdates();
                         editProfileButton.setText("Edit Profile");
+                        isEditing = false;
                     }
                 }
             });
             this.headerImage.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
-                    getImg(headerImgSelect);
-
+                    if (isEditing) {
+                        getImg(headerImgSelect);
+                    }
                 }
             });
             this.profilePicture.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
-                    getImg(profPicSelect);
-
+                    if (isEditing) {
+                        getImg(profPicSelect);
+                    }
                 }
             });
-            this.headerImage.setEnabled(true);
-            this.profilePicture.setEnabled(true);
         } else {
             this.sendMsgButton.setEnabled(true);
             this.sendMsgButton.setVisibility(View.VISIBLE);
@@ -245,8 +272,12 @@ public class ProfileFragment extends Fragment {
         uploadProfilePic();
         uploadHeaderImage();
         Map<String, Object> updates = new HashMap<>();
-        updates.put("header_image", headerUri.toString());
-        updates.put("profile_pic", profPicUri.toString());
+        if (headerUri != null) {
+            updates.put("header_image", headerUri.toString());
+        }
+        if (profPicUri != null) {
+            updates.put("profile_pic", profPicUri.toString());
+        }
         updates.put("username", usernameView.getText().toString());
         updates.put("user_bio", userBioView.getText().toString());
         db.getReference("users").child(currUser.getUid()).updateChildren(updates).addOnCompleteListener(new OnCompleteListener<Void>() {
@@ -269,10 +300,6 @@ public class ProfileFragment extends Fragment {
             public void onDataChange(@NonNull DataSnapshot snapshot) {
                 for (DataSnapshot dataSnapshot : snapshot.getChildren()) {
                     User user = dataSnapshot.getValue(User.class);
-                    Toast.makeText(getActivity(), user.toString(), Toast.LENGTH_LONG).show();
-
-                    //Boolean t = user.getHeader_image() == null; true
-                    //Toast.makeText(getActivity(), t.toString(), Toast.LENGTH_LONG).show();
                     if (user.getHeader_image() != null) {
                         Toast.makeText(getActivity(), "HERE", Toast.LENGTH_LONG).show();
 
@@ -324,14 +351,19 @@ public class ProfileFragment extends Fragment {
 
         FirebaseStorage storage = FirebaseStorage.getInstance();
         //StorageReference imgRef = storage.getReferenceFromUrl(headerImgUri);
-        StorageReference imgRef = storage.getReference().child(profUid);
+        StorageReference imgRef = storage.getReference().child("header_images").child(profUid);
         imgRef.getDownloadUrl().addOnSuccessListener(new OnSuccessListener<Uri>() {
             @Override
             public void onSuccess(Uri uri) {
 
                 headerUri = uri;
-                Glide.with(getContext()).load(uri).skipMemoryCache(true).into(headerImage);
-                headerImage.setImageURI(uri);
+                try {
+                    Picasso.get().load(uri).resize(headerImage.getWidth(), headerImage.getHeight()).centerCrop().into(headerImage);
+                } catch (IllegalArgumentException e) {
+                    Log.d("ERROR IMG", "header" + " " + headerImage.getWidth() + " " + headerImage.getHeight());
+                }
+                //Glide.with(getContext()).load(uri).skipMemoryCache(true).into(headerImage);
+                //headerImage.setImageURI(uri);
             }
         });
     }
@@ -339,18 +371,27 @@ public class ProfileFragment extends Fragment {
     private void loadProfPic(String profPictureUri) {
         FirebaseStorage storage = FirebaseStorage.getInstance();
         //StorageReference imgRef = storage.getReferenceFromUrl(profPictureUri);
-        StorageReference imgRef = storage.getReference().child(profUid);
+        StorageReference imgRef = storage.getReference().child("profile_pics").child(profUid);
 
         imgRef.getDownloadUrl().addOnSuccessListener(new OnSuccessListener<Uri>() {
             @Override
             public void onSuccess(Uri uri) {
                 profPicUri = uri;
-                Glide.with(getContext()).load(uri).into(profilePicture);
+                try {
+                    Picasso.get().load(uri).resize(profilePicture.getWidth(), profilePicture.getHeight()).centerCrop().into(profilePicture);
+                } catch (IllegalArgumentException e) {
+                    Log.d("ERROR IMG", "prof pic" + " " + profilePicture.getWidth() + " " + profilePicture.getHeight());
+                }
+               // Picasso.get().load(uri).resize(profilePicture.getWidth(), profilePicture.getHeight()).centerCrop().into(profilePicture);
+                //Glide.with(getContext()).load(uri).into(profilePicture);
             }
         });
     }
 
     private void uploadProfilePic() {
+        if (((BitmapDrawable) profilePicture.getDrawable()).getBitmap() == null) {
+            return;
+        }
         StorageReference storageReference = FirebaseStorage.getInstance().getReference();
         String imgPath = "profile_pics/" + profUid;
         StorageReference profPicStorageRef = storageReference.child(imgPath);
@@ -387,13 +428,15 @@ public class ProfileFragment extends Fragment {
     }
 
     private void uploadHeaderImage() {
+        if (((BitmapDrawable) headerImage.getDrawable()) == null) {
+            return;
+        }
         StorageReference storageReference = FirebaseStorage.getInstance().getReference();
         String imgPath = "header_images/" + profUid;
         StorageReference headerStorageRef = storageReference.child(imgPath);
         headerStorageRef.delete();
         //Boolean t = profPicStorageRef == null;
         //Toast.makeText(getActivity(), t.toString(), Toast.LENGTH_LONG).show();
-
         Bitmap bitmap = ((BitmapDrawable) headerImage.getDrawable()).getBitmap();
         ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
         bitmap.compress(Bitmap.CompressFormat.JPEG, 100, outputStream);
